Removing line with {{ due to blocking github pages genertion
The line contained {{ which is a control character for liquid
templates used by github pages that generated the docs site. With
that in the page and not being closed an error about it not being
closed occured and github pages stopped updating.

Removing the sentence with the usage of that to remove the error.