add go.mod file
Since a recent release, 2.19.0, mentioned bumping the major version of
this module, I thought I would prepare this change before that happens.

This change bumps the semantic import version to
github.com/Masterminds/sprig/v3, and will require an accompanying v3.0.0
tag. It is not suggested that this change merge before you are done with
the v2.x.x line.

This change is also important because in go1.13 incompatible packages are
being deprecated. That is those with v2+ tags, without semantic import
versions, that show up as vx.y.z+incompatbile. See the release notes
https://golang.org/doc/go1.13#version-validation for more details.

Fixes #145