Create all.md
This file is a one-pager for all the functions documented in the docs.
To create this I concatenated all the other markdown files together as
markdown has no way of including other pages.

We need to have a good way of updating this, maybe we can have a script
that is run each time documentation is changed.